# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
        "/api/getForm/": {
            "get": {
                "tags": [
                    "form"
                ],
                "summary": "getForm",
                "description": "getForm Endpoint",
                "operationId": "getForm",
                "parameters": [
                    {
                        "name": "step",
                        "in": "query",
                        "description": "step for form",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "form data",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Успех",
                                        "value": {
                                            "status": "success",
                                            "result": {
                                                "formData": [],
                                                "step": 1,
                                                "type": "needRequired|allowedNewStep|addedNewFields"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/saveForm/": {
            "post": {
                "tags": [
                    "form"
                ],
                "summary": "saveForm",
                "description": "saveForm Endpoint",
                "operationId": "saveForm",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "step"
                                ],
                                "properties": {
                                    "step": {
                                        "type": "number"
                                    },
                                    "formData": {
                                        "type": "json"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "save form",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Успех",
                                        "value": {
                                            "status": "success",
                                            "result": {
                                                "step": 1,
                                                "type": "needRequired|allowedNewStep|addedNewFields"
                                            }
                                        }
                                    },
                                    "result error": {
                                        "summary": "Нехватка полей",
                                        "value": {
                                            "status": "error",
                                            "error": "Поле step обязательна для заполнения"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/saveFile/": {
            "post": {
                "tags": [
                    "form"
                ],
                "summary": "saveFile",
                "description": "saveFile Endpoint",
                "operationId": "saveFile",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file[]"
                                ],
                                "properties": {
                                    "file[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "file"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "file info",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Успех",
                                        "value": {
                                            "status": "success",
                                            "resFile": "url file"
                                        }
                                    },
                                    "result error": {
                                        "summary": "Ошибка формирования файла",
                                        "value": {
                                            "status": "error",
                                            "error": "Текст"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/saveUserImg/": {
            "post": {
                "tags": [
                    "form"
                ],
                "summary": "saveUserImg",
                "description": "saveUserImg Endpoint",
                "operationId": "saveUserImg",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "file": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "file  info",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Успех",
                                        "value": {
                                            "status": "success",
                                            "resFile": "url file"
                                        }
                                    },
                                    "error": {
                                        "summary": "Нехватка полей",
                                        "value": {
                                            "status": "error",
                                            "error": "Ничего не загружено"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/checkPin/": {
            "post": {
                "tags": [
                    "Personal area"
                ],
                "summary": "checkPin",
                "description": "checkPin Endpoint",
                "operationId": "checkPin",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "pin"
                                ],
                                "properties": {
                                    "pin": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "check pin code",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Успех",
                                        "value": {
                                            "status": "success",
                                            "result": {
                                                "token": {
                                                    "token_type": "Bearer",
                                                    "expires_in": "числовое значение в секундах время жизни access_token",
                                                    "access_token": "токен доступа",
                                                    "refresh_token": "токен восстановления access_token"
                                                }
                                            }
                                        }
                                    },
                                    "result check pin error": {
                                        "summary": "Неверный пин",
                                        "value": {
                                            "status": "error"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "417": {
                        "description": "pin is empty",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result pin": {
                                        "summary": "Ошибка кода",
                                        "value": {
                                            "status": "error",
                                            "error": "Поле пин обязательна для заполнения"
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://preprod.marriator-api.fivecorners.ru/oauth/authorize",
                        "tokenUrl": "http://preprod.marriator-api.fivecorners.ru/oauth/token",
                        "refreshUrl": "http://preprod.marriator-api.fivecorners.ru/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    }
}